requirements.txt
flask
openai-whisper
torch
transformers
nltk
soundfile
ffmpeg-python
